requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pandas==2.1.0
openpyxl==3.1.2
WeasyPrint==59.0
WTForms==3.0.1
passlib==1.7.4
email-validator==2.0.0.post2
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
PyJWT==2.8.0
requests==2.31.0